refactor(config): 将配置项迁移至 Mods 配置界面并重构颜色映射加载
- 数值/布尔类配置项从客户端命令迁移到 Mods 列表的配置界面，
  使用 IConfigScreenFactory 提供 GUI 入口
- 客户端命令仅保留颜色映射重载功能，因 damagerender-damage-color.json
  是自由格式 JSON，ModConfigSpec 无法支持，命令成为其唯一重载入口
- 颜色映射改用 Codec + JsonOps 进行序列化与解析，替代手写 JSON 读写逻辑
- MERGE_MAX_AGE 类型从 Float 调整为 Double 以保持一致
- 新增 zh_cn.json 语言文件并同步 en_us.json

+++ src/main/java/net/yiran/damagerender/ClientConfig.java
@@ -12,7 +12,7 @@ public class ClientConfig {
     public static final ModConfigSpec.ConfigValue<Double> MIN_VALUE_DISPLAY;
     public static final ModConfigSpec.ConfigValue<Boolean> ENABLE_COMBINE_STRING;
     public static final ModConfigSpec.ConfigValue<Double> MERGE_DISTANCE_SQ;
-    public static final ModConfigSpec.ConfigValue<Float> MERGE_MAX_AGE;
+    public static final ModConfigSpec.ConfigValue<Double> MERGE_MAX_AGE;
 
     static {
 
@@ -33,7 +33,7 @@ public class ClientConfig {
                 .define("mergeDistanceSQ", 1.5);
         MERGE_MAX_AGE = BUILDER
                 .comment("合并生成在多少tick以内的文字")
-                .define("mergeMaxAge", 40.0f);
+                .define("mergeMaxAge", 40.0);
         SPEC = BUILDER.build();
 
     }
